Codigo/ai/stt/archieve/testes/scripts/requirements.txt
nvidia-cublas-cu12
nvidia-cudnn-cu12
nvidia-curand-cu12
numpy
faster-whisper
pyannote.audio
sounddevice
numpy
